♻️ Refactor: Sprite 이미지 (#237)
