+++ src/components/Header/TopBasicNavHeader.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRecoilValue } from 'recoil';
 import HeaderContainer from './HeaderContainer';
 import BackButton from './BackButton';
-import GlovalSprite from '../../assets/sprite/GlovalSprite';
+import GlobalSprite from '../../assets/sprite/GlobalSprite';
 import { isDarkModeState } from '../../atoms/StylesAtom';
 
 const TopBasicNavHeader = ({ onButtonClick }) => {
@@ -11,7 +11,7 @@ const TopBasicNavHeader = ({ onButtonClick }) => {
     <HeaderContainer>
       <BackButton />
       <button onClick={onButtonClick}>
-        <GlovalSprite id={isDarkMode ? 's-icon-more-vertical-dark' : 's-icon-more-vertical'} color={'white'} />
+        <GlobalSprite id={isDarkMode ? 's-icon-more-vertical-dark' : 's-icon-more-vertical'} color={'white'} />
       </button>
     </HeaderContainer>
   );
