+++ src/pages/LoginPage/SnsLoginPage/SnsLoginPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import SnsLogo from '../../../assets/images/login-logo.svg';
 import SnsLogoDark from '../../../assets/images/login-logo-dark.svg'
 import { Link } from 'react-router-dom';
-import GlovalSprite from '../../../assets/sprite/GlovalSprite';
+import GlobalSprite from '../../../assets/sprite/GlobalSprite';
 import WhiteTitle from '../../../assets/images/main-title(w).svg';
 
 import {
@@ -30,15 +30,15 @@ const SnsLoginPage = () => {
         </Logo>
         <LoginSection>
           <button className="kakao-login">
-            <GlovalSprite id="message-circle" />
+            <GlobalSprite id="message-circle" />
             카카오톡 계정으로 로그인
           </button>
           <button className="google-login">
-            <GlovalSprite id="google-g-logo-1" />
+            <GlobalSprite id="google-g-logo-1" />
             구글 계정으로 로그인
           </button>
           <button className="facebook-login">
-            <GlovalSprite id="facebook" />
+            <GlobalSprite id="facebook" />
             페이스북 계정으로 로그인
           </button>
           <SnsLoginLink>
